paciente rotas feitas

+++ src/main/java/org/prontuario/services/PacienteService.java
@@ -1,11 +1,15 @@
 package org.prontuario.services;
 
+import org.prontuario.exceptions.CpfJaCadastradoException;
+import org.prontuario.exceptions.IdNaoLocalizado;
+import org.prontuario.exceptions.NameNaoLocalizado;
 import org.prontuario.models.Paciente;
 import org.prontuario.repositories.PacienteRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
+import java.util.Optional;
 
 
 //Inserir Regra de negocio aqui
@@ -15,12 +19,16 @@ public class PacienteService {
     private final PacienteRepository pacienteRepository;
 
     //Constructor
+    @Autowired
     public PacienteService(PacienteRepository pacienteRepository) {
         this.pacienteRepository = pacienteRepository;
     }
 
     //CREATE
     public Paciente savePaciente(Paciente paciente){
+        if (pacienteRepository.existsByCpf(paciente.getCpf())){
+            throw new CpfJaCadastradoException("Já existe um paciente com esse CPF.");
+        }
         return pacienteRepository.save(paciente);
     }
 
@@ -29,47 +37,70 @@ public class PacienteService {
         return pacienteRepository.findAll();
     }
 
+    //READ(NAME)
+    public List<Paciente> getPacienteByName(String name){
+
+        List<Paciente> pacientes = pacienteRepository.findByNomeCompleto(name);
+
+        if (pacientes.isEmpty()){
+            throw new NameNaoLocalizado("Paciente "+name+" não foi localizado!");
+        }
+
+
+        return pacientes;
+    }
+
     //READ(ID)
     public Paciente getPacienteById(Long id){
-        return pacienteRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Paciente no encontrado"));
+
+        return pacienteRepository.findById(id).orElseThrow(() -> new IdNaoLocalizado("Paciente não encontrado"));
     }
 
     //UPDATE
-    public void updatePaciente(Long id, Paciente pacienteAtualizado) {
+    public Paciente updatePaciente(Long id, Paciente pacienteAtualizado) {
 
-//        Paciente pacienteExistente = getPacienteById(id);
-//        pacienteExistente.setNomeCompleto(pacienteAtualizado.getNomeCompleto());
-//        pacienteExistente.setDataNascimento(pacienteAtualizado.getDataNascimento());
-//        pacienteExistente.setGenero(pacienteAtualizado.getGenero());
-//        pacienteExistente.setEstadoCivil(pacienteAtualizado.getEstadoCivil());
-//        pacienteExistente.setRg(pacienteAtualizado.getRg());
-//        pacienteExistente.setCpf(pacienteAtualizado.getCpf());
-//
-//            pacienteExistente.setEndereco(pacienteAtualizado.getEndereco());
-//            pacienteExistente.setNumero(pacienteAtualizado.getNumero());
-//            pacienteExistente.setBairro(pacienteAtualizado.getBairro());
-//            pacienteExistente.setCep(pacienteAtualizado.getCep());
-//            pacienteExistente.setCidade(pacienteAtualizado.getCidade());
-//            pacienteExistente.setEstado(pacienteAtualizado.getEstado());
-//
-//            pacienteExistente.setTelefone(pacienteAtualizado.getTelefone());
-//            pacienteExistente.setEmail(pacienteAtualizado.getEmail());
-//
-//
-//            pacienteExistente.setAlergias(pacienteAtualizado.getAlergias());
-//            pacienteExistente.setDoencas(pacienteAtualizado.getDoencas());
-//            pacienteExistente.setMedicamentos(pacienteAtualizado.getMedicamentos());
-//            pacienteExistente.setHistoricoMedico(pacienteAtualizado.getHistoricoMedico());
-//            pacienteExistente.setTipoSanguineo(pacienteAtualizado.getTipoSanguineo());
-//
-//            pacienteExistente.setAltura(pacienteAtualizado.getAltura());
-//            pacienteExistente.setPeso(pacienteAtualizado.getPeso());
-//
-//            return pacienteRepository.save(pacienteExistente);
+        Optional<Paciente> paciente = Optional.ofNullable(getPacienteById(id));
+
+
+        if (paciente.isEmpty()) {
+            throw new IdNaoLocalizado("Paciente não encontrado para o ID: " + id);
+        }
+
+
+        Paciente pacienteExistente = paciente.get();
+
+
+        pacienteExistente.setNomeCompleto(pacienteAtualizado.getNomeCompleto());
+        pacienteExistente.setGenero(pacienteAtualizado.getGenero());
+        pacienteExistente.setEstadoCivil(pacienteAtualizado.getEstadoCivil());
+        pacienteExistente.setEmail(pacienteAtualizado.getEmail());
+        pacienteExistente.setTelefone(pacienteAtualizado.getTelefone());
+        pacienteExistente.setCep(pacienteAtualizado.getCep());
+        pacienteExistente.setEndereco(pacienteAtualizado.getEndereco());
+        pacienteExistente.setBairro(pacienteAtualizado.getBairro());
+        pacienteExistente.setCidade(pacienteAtualizado.getCidade());
+        pacienteExistente.setEstado(pacienteAtualizado.getEstado());
+        pacienteExistente.setNumero(pacienteAtualizado.getNumero());
+        pacienteExistente.setDataNascimento(pacienteAtualizado.getDataNascimento());
+        pacienteExistente.setCpf(pacienteAtualizado.getCpf());
+        pacienteExistente.setAlergias(pacienteAtualizado.getAlergias());
+        pacienteExistente.setDoencas(pacienteAtualizado.getDoencas());
+        pacienteExistente.setMedicamentos(pacienteAtualizado.getMedicamentos());
+        pacienteExistente.setHistoricoMedico(pacienteAtualizado.getHistoricoMedico());
+        pacienteExistente.setTipoSanguineo(pacienteAtualizado.getTipoSanguineo());
+        pacienteExistente.setPeso(pacienteAtualizado.getPeso());
+        pacienteExistente.setAltura(pacienteAtualizado.getAltura());
+
+
+        return pacienteRepository.save(pacienteExistente);
     }
 
     //DELETE
     public void deletePaciente(Long id){
-        pacienteRepository.deleteById(id);
+        Optional<Paciente> paciente = Optional.ofNullable(getPacienteById(id));
+
+        if (paciente.isPresent()) {
+            pacienteRepository.deleteById(id);
+        }
     }
 }
